# De-mutex GlyphAtlas; expose glyph information to workers via message.