--- a/src/mbgl/layout/symbol_layout.hpp
+++ b/src/mbgl/layout/symbol_layout.hpp
@@ -16,8 +16,6 @@ namespace mbgl {
 
 class GeometryTileLayer;
 class CollisionTile;
-class SpriteAtlas;
-class GlyphAtlas;
 class SymbolBucket;
 class Anchor;
 
@@ -32,12 +30,11 @@ public:
     SymbolLayout(const style::BucketParameters&,
                  const std::vector<const style::Layer*>&,
                  const GeometryTileLayer&,
-                 SpriteAtlas&);
+                 IconDependencies&,
+                 uintptr_t,
+                 GlyphDependencies&);
 
-    bool canPrepare(GlyphAtlas&);
-
-    void prepare(uintptr_t tileUID,
-                 GlyphAtlas&);
+    void prepare(const GlyphPositionMap& glyphs, const IconAtlasMap& icons);
 
     std::unique_ptr<SymbolBucket> place(CollisionTile&);
 
@@ -45,7 +42,6 @@ public:
 
     enum State {
         Pending,  // Waiting for the necessary glyphs or icons to be available.
-        Prepared, // The potential positions of text and icons have been determined.
         Placed    // The final positions have been determined, taking into account prior layers.
     };
 
@@ -80,8 +76,8 @@ private:
 
     style::SymbolLayoutProperties::PossiblyEvaluated layout;
     float textMaxSize;
-
-    SpriteAtlas& spriteAtlas;
+    
+    uintptr_t spriteAtlasMapIndex; // Actually a pointer to the SpriteAtlas for this symbol's layer, but don't use it from worker threads!
 
     const uint32_t tileSize;
     const float tilePixelRatio;
@@ -89,7 +85,6 @@ private:
     bool sdfIcons = false;
     bool iconsNeedLinear = false;
 
-    GlyphRangeSet ranges;
     std::vector<SymbolInstance> symbolInstances;
     std::vector<SymbolFeature> features;