--- a/src/mbgl/style/layers/symbol_layer_impl.cpp
+++ b/src/mbgl/style/layers/symbol_layer_impl.cpp
@@ -36,11 +36,15 @@ std::unique_ptr<Bucket> SymbolLayer::Impl::createBucket(const BucketParameters&,
 
 std::unique_ptr<SymbolLayout> SymbolLayer::Impl::createLayout(const BucketParameters& parameters,
                                                               const std::vector<const Layer*>& group,
-                                                              const GeometryTileLayer& layer) const {
+                                                              const GeometryTileLayer& layer,
+                                                              GlyphDependencies& glyphDependencies,
+                                                              IconDependencyMap& iconDependencyMap) const {
     return std::make_unique<SymbolLayout>(parameters,
                                           group,
                                           layer,
-                                          *spriteAtlas);
+                                          iconDependencyMap[spriteAtlas],
+                                          (uintptr_t)spriteAtlas,
+                                          glyphDependencies);
 }
 
 IconPaintProperties::Evaluated SymbolLayer::Impl::iconPaintProperties() const {