--- a/src/mbgl/style/source_impl.cpp
+++ b/src/mbgl/style/source_impl.cpp
@@ -191,12 +191,6 @@ void Source::Impl::removeTiles() {
     }
 }
 
-void Source::Impl::updateSymbolDependentTiles() {
-    for (auto& pair : tiles) {
-        pair.second->symbolDependenciesChanged();
-    }
-}
-
 void Source::Impl::reloadTiles() {
     cache.clear();