--- a/src/mbgl/style/style.cpp
+++ b/src/mbgl/style/style.cpp
@@ -274,12 +274,6 @@ void Style::updateTiles(const UpdateParameters& parameters) {
     }
 }
 
-void Style::updateSymbolDependentTiles() {
-    for (const auto& source : sources) {
-        source->baseImpl->updateSymbolDependentTiles();
-    }
-}
-
 void Style::relayout() {
     for (const auto& sourceID : updateBatch.sourceIDs) {
         Source* source = getSource(sourceID);
@@ -575,7 +569,6 @@ void Style::setObserver(style::Observer* observer_) {
 
 void Style::onGlyphsLoaded(const FontStack& fontStack, const GlyphRange& glyphRange) {
     observer->onGlyphsLoaded(fontStack, glyphRange);
-    updateSymbolDependentTiles();
 }
 
 void Style::onGlyphsError(const FontStack& fontStack, const GlyphRange& glyphRange, std::exception_ptr error) {
@@ -626,7 +619,6 @@ void Style::onTileError(Source& source, const OverscaledTileID& tileID, std::exc
 void Style::onSpriteLoaded() {
     observer->onSpriteLoaded();
     observer->onUpdate(Update::Repaint); // For *-pattern properties.
-    updateSymbolDependentTiles();
 }
 
 void Style::onSpriteError(std::exception_ptr error) {