--- a/src/mbgl/tile/tile.hpp
+++ b/src/mbgl/tile/tile.hpp
@@ -50,7 +50,6 @@ public:
     virtual Bucket* getBucket(const style::Layer&) = 0;
 
     virtual void setPlacementConfig(const PlacementConfig&) {}
-    virtual void symbolDependenciesChanged() {};
     virtual void redoLayout() {}
 
     virtual void queryRenderedFeatures(