… src/mbgl/text/glyph_atlas.cpp
… src/mbgl/text/glyph_atlas.hpp
… src/mbgl/text/glyph_set.cpp
… src/mbgl/text/glyph_set.hpp
… src/mbgl/text/shaping.cpp
… src/mbgl/tile/geometry_tile.cpp
… src/mbgl/tile/geometry_tile.hpp
… src/mbgl/tile/geometry_tile_worker.cpp
… src/mbgl/tile/geometry_tile_worker.hpp
… src/mbgl/util/exclusive.hpp
… test/sprite/sprite_atlas.test.cpp
… test/text/glyph_atlas.test.cpp
… test/text/glyph_pbf.test.cpp