This PR ports the gl-js `GlyphAtlas` architecture (roughly) to gl-native. Instead of using a mutex-protected `GlyphAtlas` shared across all worker threads, we restrict access to the `GlyphAtlas` to the main thread, and use a message passing interface for worker threads to request glyph information from the atlas.

Working on Harfbuzz/Freetype code motivated me to make these changes -- as the shaping code grew more complex/expensive, holding onto the `GlyphAtlas` mutex became more of a point of resource contention. Also, reasoning about mutexes is hard.

I haven't done any performance benchmarking on this PR yet. In theory it should reduce contention, but it also introduces some extra copying and the extra round trip of always having to send a message to the main thread for glyphs even if they're already locally available.

I'm putting up the PR early to get feedback on the approach. If the approach looks good, we can also use it for `SpriteAtlas`, and at that point the worker threads should (?) be running independent of shared state the same way they do in gl-js. If we're doing the same changes for both atlases, we could also update all the "symbol dependency" code to share one common pathway. Right now this new pathway is just kind of glued on to the `symbolDependenciesChanged` code from #6928.

I started out by porting gl-js callbacks more-or-less directly to C++ lambdas, but tracking ownership quickly got... complicated. Instead, I made `GlyphAtlas` directly responsible for tracking who needs to get sent what data when glyphs become available. There's more code than I'd like added to `GlyphAtlas`, but hopefully it's a simplification overall.

/cc @kkaefer @jfirebaugh @lucaswoj @mollymerp