@ChrisLoer 2017-03-14
I'm moving forward with making matching changes in `SpriteAtlas`, which allows us to get rid of most of the "symbolDependenciesChanged" logic.

One performance consideration I didn't consider until now is that we won't be able to start on `prepare` calls until all glyphs/sprites are available. Before, if one `SymbolLayout` depended on a network request, we could go ahead and do `prepare` calls for everything else while we waited for the reply -- and then when the network request finished, we'd have less work to do to get the tile ready.

If we think we need to keep the incremental preparation capability, I'll need to change the `onGlyphsAvailable` logic to be more granular. If we think it's not that important, then we can simplify `GeometryTileWorker` even more (`attemptPlacement` just becomes `place`, we stop having to keep track of `SymbolLayout::state`, etc.).

/cc @mourner