@jfirebaugh 2017-03-14
My guess is that incremental preparation is not critical. It looks like GL JS does not use it. But I don't think it's ever been benchmarked or otherwise empirically measured.

> we stop having to keep track of `SymbolLayout::state`

The other thing that `SymbolLayout::state` is used for is to avoid re-doing preparation when re-doing placement, i.e. when the map is rotated. This is probably more valuable to preserve (although again, no empirical data on this).