@jfirebaugh 2017-03-15
Pushed a few more suggested simplifications in d3c91dbaebd6cdfc19e0e7c85bb0040f91fa11fe. Further possible refactoring I see:

* Inline `GlyphAtlas::addGlyph` into `addGlyphs` (avoiding some repetitive `entries[fontStack]` lookups).
* Combine the `Entry::glyphValues` and `GlyphSet::sdfs` maps (both keyed by glyph ID). This could be a part of a larger refactor:
* Merge `GlyphSet` and `GlyphPBF` into `GlyphAtlas`. These auxiliary classes don't have a clear role. It makes more sense to merge them, and then consider how to extract `GlyphAtlas::Entries` as a standalone class instead (mainly for testing).