@ChrisLoer 2017-03-21
I do not have the ASCII diagramming skills necessary to include the updated `GeometryTileWorker` state transition diagram in the comments, but as part of the refactoring I drew up as detailed a version as I could:

![geometrytileworker](https://cloud.githubusercontent.com/assets/375121/24174127/4de679dc-0e4c-11e7-8365-a346d505d66a.png)

I'm much less familiar with the `SpriteAtlas` code than the `GlyphAtlas` code, so it would be nice to get some eyes on this from someone who knows annotations/etc. I had some open questions:

* Is it safe to assume that all access to `SpriteAtlas` is on the main thread now that's I've removed the access from `GeometryTileWorker`? It appears that way to me, but I'm not very familiar at all with annotations.
* I started development with the idea that we'd have to track individual icon dependencies, but it looks like we only really ask "is this `SpriteAtlas` loaded". If there's no value to tracking individual dependencies, I can strip out some of that code.
* `GeometryTileWorker` assumes it can be dependent on any number of sprite atlases, but it looks like usually there's only two (one for the annotation manager, one for the style). If that's true, that might allow for some simplification -- and also if we knew what sprite atlases were needed before we started layout, we could tell the worker ahead of time if they were already loaded and skip the `getIcons` round trip.