@ChrisLoer 2017-03-22
I did some basic benchmarking by running the macosapp in release mode, with the map centered over greater asia and using a style that loads `{name}` instead of `{name_en}` (so the test was designed to have lots of glyph dependencies). I recorded milliseconds from initialization of the `Map` object to the time of the last `update` call to the map, doing four runs on `master` and four runs on this branch. I also re-did the tests with the cache deleted between each run.

| | Master | De-Mutexed |
|---|---|---|
| With Cache | 1107 ms | 888 ms |
| No Cache | 1818 ms | 1717 ms | 

These numbers make me feel fairly confident that this branch isn't introducing a major performance regression, although there are definitely theoretical cases where these changes could slow us down.