@jfirebaugh 2017-03-22
The test failure might be an instance of #6863 -- it's hard to tell exactly because the icon is partially obscured by text.