@ChrisLoer 2017-03-22
And it still happens even with bigger basketballs: http://mapbox.s3.amazonaws.com/mapbox-gl-native/render-tests/10010/index.html

It kind of looks like the vertices for the lower right triangle are ending up shifted one pixel right from the vertices for the upper left triangle.