@ChrisLoer 2017-03-23
I did an experiment with the [fragment shader doing its own rounding](https://github.com/mapbox/mapbox-gl-native/commit/925c79ebc95f4e7c2e864a2f4af22fc46019237e)... the results (http://mapbox.s3.amazonaws.com/mapbox-gl-native/render-tests/10015/index.html) show lots of little changes for icons, but for the text-font/chinese test the broken behavior is totally unaffected.