@ChrisLoer 2017-03-23
On the other hand, making the symbol icon vertex shader round `gl_Position` to integer values solves the split-icon problem in text-font/chinese (at the expense of making lots of small changes all over the map). http://mapbox.s3.amazonaws.com/mapbox-gl-native/render-tests/10019/index.html